AD-HOC feat (Server): Install and minimally configure the server
In previous work the scaffolding and documentation of the repository was
created. This commit adds the first set of logic, installing the
Prometheus server (and, as a dependency, the node exporter).
Additionally the server has been minimally confingured.

== Design Notes ==

=== Variables of the style ${NAMESPACE}__${ACTION} ===

In a style initially popularised by the debops playbooks, this role
starts naming variables namespaced by the role name. This makes the
intended variable usage clear, as well as hopefully prevents collisions
between variables.