AD-HOC refactor (node exporter): Implement the node exporter
This commit extends on the previous work of implementing the server by
explicitly adding the node exporter and allowing configuration.

It additionally refactors some other cruft in the repo. Principally it
relabels the prometheus-server to be simply prometheus, as well as moves
variables from the vars/ folder to defaults/ where they can be
overriden.

It's all chunked in a single commit because Sunday.