AD-HOC refactor (install): Issue install instructions on a per OS basis
Because different operating systems have different mechanisms to install
Prometheus, it stands to reason that each operating system should
install it its own way, with only the common parts of the playbook being
non OS specific.

This will introduce some complexity as there are many systems that do
install Prometheus at the same version in the same way, but it is
thought that this can be resolved with symlinks.