AD-HOC feat (Prometheus): Add system user, configuration
Currently there is a problem in which Prometheus is unable to read the
configuration as Prometheus when installed via apt is being run as
"Prometheus" and not as root, but the configuration is owned by root.

This commit modifies binary installation to also run as "prometheus",
and the configuration to be owned by this user.